Исправлено окончание количества вакансий2
