+++ app/src/main/java/ru/practicum/android/diploma/util/Utils.kt
@@ -1,5 +1,8 @@
 package ru.practicum.android.diploma.util
 
+import android.content.Context
+import ru.practicum.android.diploma.R
+
 //константы уровня приложения
 
 const val CLICK_DEBOUNCE_DELAY = 1000L
@@ -29,8 +32,23 @@ const val ONE = 1
 const val CHECK_CONNECTION = "Проверьте подключение к интернету"
 const val ERROR_HAS_OCCURRED = "Произошла ошибка"
 const val NOT_EXIST = "Таких вакансий нет"
-
+const val FOUND_0_VACANCIES = "Найдено %1\$s вакансий"
+const val FOUND_1_VACANCIES = "Найдена %1\$s вакансия"
+const val FOUND_2_VACANCIES = "Найдено %1\$s вакансии"
 
 const val BASE_URL = "https://api.hh.ru/"
 
 const val SHARED_PREFS = "app_preferences"
+
+fun getStringCount(count: String): String {
+    val digit = count.last().toString().toInt()
+    return if (count.takeLast(2).toInt() in 10..20) {
+        FOUND_0_VACANCIES.format(count)
+    } else if (digit == 1) {
+        FOUND_1_VACANCIES.format(count)
+    } else if (digit in 2..4) {
+        FOUND_2_VACANCIES.format(count)
+    } else {
+        FOUND_0_VACANCIES.format(count)
+    }
+}
